clarity_backend/requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
asgiref==3.6.0
astroid==2.14.2
async-timeout==4.0.2
attrs==22.2.0
black==23.1.0
certifi==2022.12.7
charset-normalizer==2.0.12
click==8.1.3
coreapi==2.3.3
coreschema==0.0.4
coverage==7.2.1
crispy-bootstrap5==0.7
defusedxml==0.7.1
diff-match-patch==20200713
dill==0.3.6
Django==4.1.7
django-ckeditor==6.5.1
django-cors-headers==3.14.0
django-crispy-forms==2.0
django-extensions==3.2.1
django-import-export==3.1.0
django-jazzmin==2.6.0
django-js-asset==2.0.0
django-pytest==0.2.0
django-widget-tweaks==1.4.12
django_debug_toolbar==3.8.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-spectacular==0.25.1
drf-yasg==1.21.5
et-xmlfile==1.1.0
exceptiongroup==1.1.0
frozenlist==1.3.3
gunicorn==20.1.0
h11==0.14.0
idna==3.4
ifcfg==0.22
inflection==0.5.1
iniconfig==2.0.0
isort==5.12.0
itypes==1.2.0
Jinja2==3.1.2
jmespath==0.10.0
jsonschema==4.17.3
lazy-object-proxy==1.9.0
MarkupPy==1.14
MarkupSafe==2.1.2
mccabe==0.7.0
multidict==6.0.4
mypy-extensions==1.0.0
numpy==1.24.2
odfpy==1.4.1
openai==0.26.5
openpyxl==3.1.1
packaging==23.0
pandas==1.3.5
pathspec==0.11.0
Pillow==9.4.0
platformdirs==3.0.0
pluggy==1.0.0
psycopg2-binary==2.9.1
PyJWT==2.3.0
pylint==2.16.2
pyrsistent==0.19.3
pytest==7.2.1
pytest-cov==4.0.0
pytest-django==4.5.2
python-dateutil==2.8.2
pytz==2021.3
PyYAML==6.0
requests==2.27.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
six==1.16.0
sqlparse==0.4.2
tablib==3.3.0
tomli==2.0.1
tomlkit==0.11.6
tqdm==4.64.1
typing_extensions==4.1.1
uritemplate==4.1.1
urllib3==1.26.8
whitenoise==5.3.0
wrapt==1.15.0
xlrd==2.0.1
xlwt==1.3.0
yarl==1.8.2
